Implement more of the protocol (#79)
* Write Clientbound::Respawn

* Fix Slot.empty? numeric id

* Fix old yaw/pitch_deg

* update health/food/saturation

* gamemode is UInt8

* Use server-provided username

* Use Chat object where applicable

closes #31

+++ src/rosegold/world/player_list.cr
@@ -5,7 +5,7 @@ module Rosegold::PlayerList
     getter uuid : UUID
     property name : String?
     property properties = [] of Property
-    property gamemode : UInt32? # TODO allow negative
+    property gamemode : Int8?
     property ping : UInt32?
     property display_name : Rosegold::Chat?
